frotz-issue-44: Replaced type "ulong" with type "unsigned long" to enable support for Apple MacOS Sierra 10.12 plus MacPorts 2.4.1.

+++ src/curses/ux_blorb.h
@@ -22,7 +22,7 @@ typedef struct sampledata_struct {
 typedef struct {
     bb_result_t bbres;
     // ulong type;
-    uint32_t type;
+    unsigned long type;
     FILE *fp;
 } myresource;
